% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{compute_epe}
\alias{compute_epe}
\title{Generalization Error}
\usage{
compute_epe(
  data,
  randx,
  testn,
  fit,
  randy,
  error = c("squared", "absolute"),
  distribution = FALSE,
  mod,
  formula,
  ...
)
}
\arguments{
\item{data}{MC data set generated by \code{\link{mc_data}}.}

\item{randx}{Random sample generator function for \code{x}. By default, \link[stats]{rnorm}}

\item{testn}{Test sample size}

\item{fit}{True model function with \code{x}-named argument.}

\item{randy}{Random sample generator function for error term. By default, \link[stats]{rnorm}}

\item{error}{Choice of loss function. See \code{\link{loss}}.}

\item{distribution}{return the error for each MC sample? \code{FALSE} by default. If \code{TRUE}, it gives the \code{data.table}.}

\item{mod}{Model function.}

\item{formula}{an object of class \link[stats]{formula}.}

\item{...}{Additional arguments for \code{mod}. If you wand argument for \code{randx} or \code{randy}, define one.}
}
\value{
Expected test error
}
\description{
This function computes expected test error in Monte Carlo simultion situation.
}
\details{
Given MC samples, compute test error using independent test set and average.
}
\references{
Hastie, T., Tibshirani, R.,, Friedman, J. (2001). \emph{The Elements of Statistical Learning}. New York, NY, USA: Springer New York Inc..
}
