% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{gen_y}
\alias{gen_y}
\title{Generate Response}
\usage{
gen_y(
  data,
  fit,
  rand,
  mcname = "mc",
  xname = "x",
  yname = "y",
  fitname = "fx",
  fit_col = TRUE,
  ...
)
}
\arguments{
\item{data}{MC data set generated by \code{\link{mc_data}}.}

\item{fit}{True model function with \code{xname} argument.}

\item{rand}{Random sample generator function for error term. By default, \link[stats]{rnorm}}

\item{mcname}{column name of the MC sample. By default, \code{"mc"}.}

\item{xname}{column name of the data. By default, \code{"x"}.}

\item{yname}{column name of the response. By default, \code{"y"}.}

\item{fitname}{column name of the true fit. By default, \code{"fx"}.}

\item{fit_col}{Add the true function value column? \code{TRUE} by default.}

\item{...}{additional arguments for \code{rand}}
}
\value{
Add two columns named \code{fx}, a true function value and \code{y}, error added.
}
\description{
This function generates the response variable using true model assumed.
}
