% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{\%m\%}
\alias{\%m\%}
\title{Loss operation}
\usage{
lhs \%m\% rhs
}
\arguments{
\item{lhs}{numeric or character}

\item{rhs}{numeric or character}
}
\value{
For numeric, compute division. Otherwise, misclassification.
}
\description{
Binary operation method for character
}
\details{
This function can deal with both regression and classification problem.
}
