% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss}
\alias{loss}
\title{Loss function}
\usage{
loss(lhs, rhs, error = c("squared", "absolute"))
}
\arguments{
\item{lhs}{numeric or character}

\item{rhs}{numeric or character}

\item{error}{choice of error. See 'Details'.}
}
\value{
Error between numeric or class.
}
\description{
Various loss functions
}
\details{
This function can compute squared error (\code{"squared"}) and absolute error (\code{"absolute"}) now.
}
