% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc_data}
\alias{mc_data}
\title{Generate Monte Carlo Samples}
\usage{
mc_data(rand, N = 100, M = 1000, mcname = "mc", xname = "x", char = "s", ...)
}
\arguments{
\item{rand}{Random sample generator function with parameter n}

\item{N}{Sample size in each MC sample}

\item{M}{Number of MC samples}

\item{mcname}{column name of the MC sample. By default, \code{"mc"}.}

\item{xname}{column name of the data. By default, \code{"x"}.}

\item{char}{Indicator for MC sample, "s" by default. Distinguished by suffix number.}

\item{...}{additional arguments for \code{rand}}
}
\value{
\code{data.table} with two columns. \code{x} is generated samples and \code{mc} indicates the MC sample.
}
\description{
This function generates \code{\link[data.table]{data.table}} which consists of Monte Carlo sets.
}
\details{
This function prepares for \code{data.table} group operation.
By grouping \code{mc} column, MC simulation can be easily done.
}
