% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{gen_y}
\alias{gen_y}
\title{Generate Response}
\usage{
gen_y(data, model, rand, ...)
}
\arguments{
\item{data}{MC data set generated by \code{\link{mc_data}}.}

\item{model}{True model function with \code{x}-named argument.}

\item{rand}{Random sample generator function for error term. By default, \link[stats]{rnorm}}
}
\value{
Add two columns named \code{fx}, a true function value and \code{y}, error added.
}
\description{
This function generates the response variable using true model assumed.
}
