% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{pred_dt}
\alias{pred_dt}
\title{Train and Predict a Model in each MC Sample}
\usage{
pred_dt(data, mod, formula, ...)
}
\arguments{
\item{data}{MC data set form of \code{\link{mc_data}}.}

\item{mod}{Model function.}

\item{formula}{an object of class \link[stats]{formula}.}

\item{...}{Additional arguments for \code{mod}}
}
\value{
The function adds a column named \code{pred} of which values are the fitted values from \code{mod}
}
\description{
This function produces fitted values in each MC sample.
}
\details{
\code{mod} should have \code{formula} argument yet.
It uses predict.mod() function. Some class has an option called \code{type}. For example, \link[stats]{predict.glm}.
This function uses \code{type = "response"}.
Some model functions require \code{x} and \code{y}, so I will add that feature.
}
