% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youngtool-package.R
\docType{package}
\name{youngtool-package}
\alias{youngtool}
\alias{youngtool-package}
\title{youngtool: Research tools for academic writing}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Reasearch tools that Young-geun uses
    This package has many functions when writing academic works.
}
\details{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

The youngtool pakcage provides a bunch of functions that can be helpful to research.
For example, it does Monte Carlo simulation using S3 [`data.table`] class structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ygeunkim}
}

}
\keyword{internal}
